+++ examples/test-site/src/components/FlowContainer/index.tsx
@@ -21,8 +21,10 @@ import withSingleAccordionVariations from './withSingleAccordionVariations';
 import withImageVariations from './withImageVariations';
 import withIframeVariations from './withIframeVariations';
 import withYouTubeVariations from './withYouTubeVariations';
+import withSocialShare from './withSocialShare';
 
 import { asFlowContainerWithMargins } from './token';
+import withListVariations from './withListVariations';
 
 // Order of includes currently dictates order in Component Picker
 // thus recommend putting more frequently used components toward top for quicker access.
@@ -32,7 +34,9 @@ const FlowContainerDefault = flow(
   withToutVariations,
   withContentfulTouts,
   withSingleAccordionVariations,
+  withListVariations,
   withIframeVariations,
+  withSocialShare,
   withYouTubeVariations,
   asFlowContainerWithMargins,
   withMandatoryCategories(['Orientation', 'Type']),