+++ packages/bodiless-components/src/index.ts
@@ -13,28 +13,31 @@
  */
 
 import Editable, { withPlaceholder, asEditable } from './Editable';
-import { asBodilessLink } from './Link';
+import { asBodilessLink, withoutLinkWhenLinkDataEmpty } from './Link';
+import type { AsBodilessLink } from './Link';
 import Image, {
   asBodilessImage,
   TImagePickerUI,
   withImagePlaceholder,
+  DropZonePlugin as ImageDropZone,
 } from './Image';
+import type { AsBodilessImage } from './Image';
 import NodeViewer from './NodeViewer';
 import withLinkToggle from './LinkToggle';
-import List, { asTestableList } from './List';
+import List from './List_DEPRECATED';
 import asTaggableItem from './Taggable/asTaggableItem';
-import withListTitle from './List/withListTitle';
-import asEditableList from './List/asEditableList';
-import asBasicSublist from './List/asBasicSublist';
-import withSublist, { withBasicSublist } from './List/withSublist';
-import withDeleteSublistOnUnwrap from './List/withDeleteSublistOnUnwrap';
+import withListTitle from './List_DEPRECATED/withListTitle';
+import asEditableList from './List_DEPRECATED/asEditableList';
+import asBasicSublist from './List_DEPRECATED/asBasicSublist';
+import withSublist, { withBasicSublist } from './List_DEPRECATED/withSublist';
+import withDeleteSublistOnUnwrap from './List_DEPRECATED/withDeleteSublistOnUnwrap';
 import {
   withMeta, withMetaStatic, withMetaHtml, withTitle,
 } from './Meta/Meta';
 import withMetaForm, { withMetaSnippet } from './Meta/withMetaForm';
 import type { FieldType as MetaFormFieldType } from './Meta/withMetaForm';
 import asBodilessHelmet from './Helmet/Helmet';
-import withEvent from './GTM/gtm';
+import withDataLayerItem, { withDefaultDataLayer, withDataLayerScript } from './GTM/gtm';
 import { withToggle, withToggleTo, withToggleButton } from './Toggle';
 import withEditPlaceholder from './Placeholder';
 import { TagButtonProps, withTagButton, useTagsAccessors } from './TagButton';
@@ -70,11 +73,15 @@ import {
   ifViewportIs,
   ifViewportIsNot,
 } from './withResponsiveToggle';
+import withBodilessLinkToggle from './withBodilessLinkToggle';
 
 export {
+  withBodilessLinkToggle,
   asBodilessLink,
+  withoutLinkWhenLinkDataEmpty,
   Image,
   asBodilessImage,
+  ImageDropZone,
   withImagePlaceholder,
   TImagePickerUI,
   Editable,
@@ -83,7 +90,6 @@ export {
   List,
   asEditableList,
   asBasicSublist,
-  asTestableList,
   withBasicSublist,
   withSublist,
   withDeleteSublistOnUnwrap,
@@ -99,7 +105,6 @@ export {
   withMetaHtml,
   asBodilessHelmet,
   withEditPlaceholder,
-  withEvent,
   withTagButton,
   TagButtonProps,
   asTaggableItem,
@@ -132,11 +137,19 @@ export {
   ifViewportIsNot,
   withMetaForm,
   withMetaSnippet,
+  withDataLayerItem,
+  withDefaultDataLayer,
+  withDataLayerScript,
 };
 
-export type {
-  FinalProps as ListProps,
-  TitleProps as ListTitleProps,
-  ListDesignableComponents,
-} from './List/types';
+export * from './Chameleon/index';
+export * from './List';
+
+export * from './Breadcrumbs';
+
 export type { MetaFormFieldType, YouTubePlayerSettings };
+
+export type {
+  AsBodilessImage,
+  AsBodilessLink,
+};