+++ packages/bodiless-core/src/Types/EditButtonTypes.tsx
@@ -35,7 +35,16 @@ export type EditButtonProps<D> = {
 };
 
 export type EditButtonOptions<P, D> = Omit<TMenuOption, 'handler'> & {
-  renderForm?: FormBodyRenderer<P, D>,
+  /**
+   * Callback to render the body of the edit form.
+   */
+  renderForm: FormBodyRenderer<P, D>,
+  /**
+   * Optional label for the context menu group to which the button will belong.
+   * If omitted, will use the button label.
+   */
+  groupLabel?: string,
+  formTitle?: string,
   /**
   * An optional function that determines if the created menu option displays "compound form".
   * If the function returns true, then "compound form" is displayed.
@@ -44,3 +53,7 @@ export type EditButtonOptions<P, D> = Omit<TMenuOption, 'handler'> & {
   */
   useCompoundForm?: () => boolean,
 } & EditDataHandler<D>;
+
+export type UseBodilessOverrides<P = any, D = any> = (
+  props: P & EditButtonProps<D>,
+) => Partial<EditButtonOptions<P, D>>;