+++ packages/bodiless-layouts/src/FlowContainer/useGetMenuOptions.tsx
@@ -15,7 +15,7 @@
 import { useCallback } from 'react';
 import { omit } from 'lodash';
 import {
-  useEditContext, useActivateOnEffect, useGetter,
+  useEditContext, useActivateOnEffect, useGetter, TMenuOption,
 } from '@bodiless/core';
 import { EditFlowContainerProps, FlowContainerItem } from './types';
 import type { FlowContainerDataHandlers, FlowContainerItemHandlers } from './model';
@@ -72,6 +72,36 @@ const useComponentSelectorActions = (
   return { insertItem, replaceItem };
 };
 
+const useCloneButton = (
+  handlers: Handlers,
+  props: EditFlowContainerProps,
+  item: FlowContainerItem,
+) => {
+  const context = useEditContext();
+  const { insertFlowContainerItem, getItems } = handlers;
+  const { maxComponents = Infinity } = props;
+  const { setId } = useActivateOnEffect();
+
+  const handler = () => {
+    const { uuid } = insertFlowContainerItem(item.type, item, item.wrapperProps);
+    setId(uuid);
+  };
+
+  const isHidden = useCallback(
+    () => !context.isEdit || getItems().length >= maxComponents, [maxComponents],
+  );
+
+  return {
+    name: 'copy-item',
+    label: 'Copy',
+    icon: 'content_copy',
+    global: false,
+    local: true,
+    handler,
+    isHidden,
+  };
+};
+
 const useDeleteButton = (
   handlers: Handlers,
   item: FlowContainerItem,
@@ -92,6 +122,8 @@ const useDeleteButton = (
     name: 'delete',
     label: 'Delete',
     icon: 'delete',
+    global: false,
+    local: true,
     handler,
     isHidden: useCallback(() => !context.isEdit, []),
   };
@@ -109,10 +141,13 @@ const useAddButton = (
   const isHidden = item
     ? useCallback(() => !context.isEdit || getItems().length >= maxComponents, [maxComponents])
     : useCallback(() => !context.isEdit || getItems().length > 0, []);
+  // @TODO For nested flow containers we'll have to give these unique names.
   const name = item ? 'add-item' : 'add';
   return {
     icon: 'add',
     label: 'Add',
+    global: false,
+    local: true,
     name,
     handler: () => componentSelectorForm(props, insertItem),
     activateContext: false,
@@ -131,6 +166,8 @@ const useSwapButton = (
     name: 'swap',
     label: 'Swap',
     icon: 'repeat',
+    global: false,
+    local: true,
     handler: () => componentSelectorForm(props, replaceItem),
     activateContext: false,
     isHidden: useCallback(() => !context.isEdit, []),
@@ -146,7 +183,7 @@ const useSwapButton = (
  */
 function useMenuOptions(props: EditFlowContainerProps) {
   const handlers = { ...useFlowContainerDataHandlers(), ...useItemHandlers() };
-  const addButton = useAddButton(handlers, withNoDesign(props));
+  const addButton: TMenuOption = useAddButton(handlers, withNoDesign(props));
   return [addButton];
 }
 
@@ -180,6 +217,7 @@ const useGetItemUseGetMenuOptions = (props: EditFlowContainerProps) => {
     const buttons = [
       // These hooks are all invoked by the flow container item (not the flow container itself).
       useAddButton(handlers, props$, item),
+      useCloneButton(handlers, props$, item),
       useSwapButton(handlers, props$, item),
       useDeleteButton(handlers, item),
     ];